authorization finish

+++ app/Http/Controllers/ArticleController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StoreArticleRequest;
 use App\Http\Requests\UpdateArticleRequest;
 use App\Models\Article;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -64,6 +65,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        Gate::authorize('update',$article);
         return view('article.edit', compact('article'));
     }
 
@@ -72,6 +74,15 @@ class ArticleController extends Controller
      */
     public function update(UpdateArticleRequest $request, Article $article)
     {
+        // if (!Gate::allows("article-update", $article)) {
+        //     return abort(403, "ပေးမသုံးဘူးကွာ၊မဆိုးနဲ့ကွာ");
+        // }
+        // if(Gate::denies('article-update',$article)){
+        //     return abort(403);
+        // }
+
+        Gate::authorize('update',$article);
+
         $article->update([
             "title" => $request->title,
             "description" => $request->description,
@@ -79,7 +90,6 @@ class ArticleController extends Controller
         ]);
 
         return redirect()->route("article.index")->with("message", $article->title . " is updated");
-
     }
 
     /**
@@ -87,8 +97,8 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize("delete",$article);
         $article->delete();
         return redirect()->route("article.index")->with("message", "Article is deleted");
-
     }
 }
